% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VAmain.r
\name{codeVA}
\alias{codeVA}
\title{Running automated method on VA data}
\usage{
codeVA(data, data.type = c("WHO", "PHMRC", "customize")[1],
  data.train = NULL, causes.train = NULL, cause.table = NULL,
  model = c("InSilicoVA", "InterVA", "Tariff")[1], Nchain = 1,
  Nsim = 10000, physician = NULL, version = "4.02", HIV = "h",
  Malaria = "h", ...)
}
\arguments{
\item{data}{}

\item{data.type}{}

\item{model}{}

\item{Nchain}{}

\item{Nsim}{}

\item{physician}{}

\item{version}{}

\item{HIV}{}

\item{Malaria}{}

\item{...}{}
}
\value{
fitted object
}
\description{
Running automated method on VA data
}
\examples{
\dontrun{
 x <- 1
}
}

