% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VAmain.r
\name{plotVA}
\alias{plotVA}
\title{Title}
\usage{
plotVA(object, top = 10, title = NULL, ...)
}
\arguments{
\item{object}{}

\item{top}{}

\item{title}{}

\item{...}{}
}
\description{
Title
}
\examples{
\dontrun{
  x <- 1
}
}

